# src/main/resources/swagger/swagger.yml
...
openapi: 3.0.0
info:
  title: Community REST API
  description: Community REST API
  version: "1.0"
...
servers:
  - url: http://localhost:8080/
  - url: https://localhost:8080/
################################################################################
#                               Components                                     #
################################################################################
components:
...
  requestBodies:
    Role:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Role"
      description: Role Status
      required: true
...
    MemberList:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MemberList"
################################################################################
#                               Parameters                                     #
################################################################################
  parameters:
    OffsetParam:
      name: offset
      in: query
      description: Offset
      required: true
      schema:
        type: integer
        format: int32
        default: -1
    LimitParam:
      name: limit
      in: query
      description: Negative number means unlimited
      required: true
      schema:
        type: integer
        format: int32
        default: -1
...
    UserIdParam:
      name: userId
      in: query
      description: User id
      required: false
      schema:
        type: string
        default: ""
...
    MemberSortParam:
      name: sort
      in: query
      description: 
        Sort
          * -role - Sort by role. Descending.
          * +name - Sort by name. Ascending.
      required: true
      schema:
        type: string
        default: "+name"
################################################################################
#                                  Schema                                      #
################################################################################
  schemas:
...
    MemberListResult:
      type: object
      properties:
        result:
          type: array
          items:
            $ref: "#/components/schemas/MemberInfoDetail"
        numFound:
          type: integer
          format: int32
...
paths:
...
  /forum/{forumId}/members:
    get:
      tags:
        - Forum
      summary: Get the members who is in or not in the forum
      operationId: getMemberListOfForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/MemberSortParam"
        - $ref: "#/components/parameters/UserIdParam"
        - name: memberInForum
          in: query
          description: >
            memberInForum:
              * true - Get the members in the forum.
              * false - Get the members in the community who are not in this forum.
          required: true
          schema:
            type: boolean
            default: true
        - name: q
          in: query
          description: query name
          required: false
          schema:
            type: string
            default: ""
        - name: isImgAvatar
          in: query
          description: 
            isimgAvatar
              * true - Get avatar.
              * false - Don't get avatar.
          required: false
          schema:
            type: boolean
            default: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListResult"
        400:
          description: Invalid parameters
    post:
      tags:
        - Forum
      summary: Add members into the forum
      operationId: addMembersIntoForum
      requestBody:
        $ref: "#/components/requestBodies/MemberList"
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: OK
        400:
          description: Invalid parameters
  /forum/{forumId}/members/application:
    get:
      tags:
        - Forum
      summary: Get the list of members need to be approved
      operationId: getReviewedMemberListOfForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicantDetail"
    post:
      tags:
        - Forum
      summary: Add the application of joining the forum.
      operationId: addMemberApplicationOfForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        $ref: "#/components/requestBodies/ApplicationDetail"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example:
                  message: Your application is sent. Please wait for the approval.
        400:
          description: Invalid parameters
  /forum/{forumId}/members/application/{applicantId}:
    patch:
      tags:
        - Forum
      summary: Review the application of joining the forum.
      operationId: reviewMemberApplicationOfForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
        - name: applicantId
          in: path
          description: Id of the member
          required: true
          schema:
            type: string
        - name: action
          in: query
          description: 
            Review action
              * approved - Approve
              * rejcted - Reject
          required: true
          schema:
            type: string
            enum: *ReviewAction
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example:
                  message: The application is approved/ rejected.
        400:
          description: Invalid parameters
  /forum/{forumId}/members/notification:
    post:
      tags:
        - Forum
      summary: Notify the members in the forum.
      operationId: notifyMemberOfForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        $ref: "#/components/requestBodies/NotificationDetail"
      responses:
        200:
          description: OK
        400:
          description: Invalid parameters
  /forum/{forumId}/members/{memberId}:
    post:
      tags:
        - Forum
      summary: Add a member into the forum
      operationId: addMemberIntoForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
        - name: memberId
          in: path
          description: Id of the member
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: OK
        400:
          description: Invalid parameters
    delete:
      tags:
        - Forum
      summary: Delete a member from the forum
      operationId: deleteMemberFromForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
        - name: memberId
          in: path
          description: Id of the member
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example:
                  message: You have left the forum.
        400:
          description: Invalid parameters
  /forum/{forumId}/role/{memberId}:
    put:
      tags:
        - Forum
      summary: Update member's role in the forum
      operationId: updatMemberRoleOfForum
      parameters:
        - name: forumId
          in: path
          description: Id of the community
          required: true
          schema:
            type: integer
            format: int32
        - name: memberId
          in: path
          description: Id of the member
          required: true
          schema:
            type: string
      requestBody:
        $ref: "#/components/requestBodies/Role"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
        400:
          description: Invalid parameters
  /forum/{forumId}:
    get:
      tags:
        - Forum
      summary: Get the forum information
      operationId: getForumInfo
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumHomePage"
        400:
          description: Invalid parameters
    put:
      tags:
        - Forum
      summary: to update forum information
      description: Used to update forum information
      operationId: updateForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatedForumData"
        description: 
          communityId is the Community id
          * name is the forum title
          * tag is the tag
          * type describes public or private
          * status describes open or delete
          * admins describe all administrators
          * members describe all members
          * forumModifiedTime describes current the forum moditied time
        required: true
      responses:
        201:
          description: Success
        400:
          description: Bad Request
        404:
          description: update Forum Error
        409:
          description: Conflict
    delete:
      tags:
        - Forum
      summary: to delete forum information
      operationId: deleteForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForumStatusInput"
        description: delete status
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example:
                  message: You have deleted the forum.
        400:
          description: Invalid parameters
  /forum/{forumId}/move:
    patch:
      tags:
        - Forum
      summary: Move forum to another community
      operationId: moveForumToAnotherCommunity
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForumMoveData"
        required: true
      responses:
        200:
          description: Success
        400:
          description: Invalid parameters
        409:
          description: Conflict
  /forum/{forumId}/topics:
    get:
      tags:
        - Forum
      summary: List the topics in the forum
      operationId: searchTopicListOfForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/TopicSortParam"
        - name: withTopping
          in: query
          description: with topping result
          required: true
          schema:
            type: boolean
            default: false
        - name: state
          in: query
          description: 
           unconcluded * concluded 
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicSearchResult"
        400:
          description: Invalid parameters
  /forum:
    post:
      tags:
        - Forum
      summary: to create forum
      description: Used to create forum information
      operationId: createForum
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForumData"
        description: 
          communityId is the Community id
          * name is the forum title
          * tag is the tag
          * type describes public or private
          * status describes open or delete
          * admins describe all administrators
          * members describe all members
        required: true
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: integer
                format: int32
        202:
          description: Accepted
        400:
          description: Bad Request
        404:
          description: Create Forum Error
...
  /forum/{forumId}/pin:
    patch:
      tags:
        - Forum
      summary: pin a forum
      operationId: updatPinOfForum
      parameters:
        - name: forumId
          in: path
          description: Id of the forum
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Pin"
        description: to pin or unpin forum
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
        400:
          description: Invalid parameters
...